Move karma safari plugin to scripts/ci-test
