+++ scripts/ci-test/karmaSafariLauncher.js
