--- tools/karma-safari-launcher/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "karma-safari-launcher",
-  "version": "1.0.0",
-  "description": "A Karma plugin to launch a Safari browser",
-  "main": "index.js"
-}
